navbar component created
