+++ pruebas/01-reading-list/diego-gonzales/src/app/app.component.html
@@ -1,16 +1,5 @@
-<header
-  class="flex flex-col md:flex-row justify-between items-center px-8 my-4 max-w-7xl mx-auto"
->
-  <a class="text-3xl text-center font-bold" routerLink="/">
-    Reading List App
-  </a>
-  <input
-    type="text"
-    placeholder="Search book"
-    class="input input-bordered hidden md:block"
-    *ngIf="showSearchBox()"
-    (input)="onInputChange($event)"
-  />
+<header class="px-8 my-4 max-w-7xl mx-auto">
+  <app-navbar />
 </header>
 
 <main class="px-4 md:px-8 pb-6 max-w-7xl mx-auto">
