+++ pruebas/01-reading-list/diego-gonzales/src/app/app.component.ts
@@ -3,30 +3,22 @@ import { CommonModule } from '@angular/common';
 import { RouterLink, RouterOutlet } from '@angular/router';
 import { BooksService } from './services/books.service';
 import { STORAGE_KEY } from './consts';
+import { NavbarComponent } from './components/shared/navbar/navbar.component';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [CommonModule, RouterOutlet, RouterLink],
+  imports: [CommonModule, RouterOutlet, RouterLink, NavbarComponent],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
 })
 export class AppComponent {
   private _booksService = inject(BooksService);
-  showSearchBox = this._booksService.showSearchBox;
 
   @HostListener('window:storage', ['$event'])
   onStorageChange(event: StorageEvent) {
     if (event.key !== STORAGE_KEY) return;
 
     this._booksService.validateStorage();
-  }
-
-  onInputChange(event: Event) {
-    const value = (event.target as HTMLInputElement).value.trim();
-
-    this._booksService.filters.mutate((filters) => {
-      filters.search = value;
-    });
   }
 }
