+++ pruebas/01-reading-list/diego-gonzales/src/app/components/shared/navbar/navbar.component.ts
@@ -1,0 +1,22 @@
+import { Component, inject } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { BooksService } from '~/services/books.service';
+
+@Component({
+  selector: 'app-navbar',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './navbar.component.html',
+})
+export class NavbarComponent {
+  private _booksService = inject(BooksService);
+  showSearchBox = this._booksService.showSearchBox;
+
+  onInputChange(event: Event) {
+    const value = (event.target as HTMLInputElement).value.trim();
+
+    this._booksService.filters.mutate((filters) => {
+      filters.search = value;
+    });
+  }
+}
